Use the real User model for the users viewset.
No-Issue
