+++ galaxy_ng/app/api/v1/viewsets/__init__.py
@@ -1,5 +1,8 @@
 from .namespaces import (
     LegacyNamespacesViewSet,
+)
+
+from .users import (
     LegacyUsersViewSet
 )
